random-number-generator/requirements.txt
numpy
matplotlib
scipy.stats